Start removing old steps and pipelines
