+++ src/zenml/__init__.py
@@ -47,10 +47,10 @@ from zenml.artifacts.artifact_config import ArtifactConfig
 from zenml.artifacts.external_artifact import ExternalArtifact
 from zenml.model.model import Model
 from zenml.model.model_version import ModelVersion # TODO: deprecate me
-from zenml.new.pipelines.pipeline_context import get_pipeline_context
-from zenml.new.pipelines.pipeline_decorator import pipeline
-from zenml.new.steps.step_decorator import step
-from zenml.new.steps.step_context import get_step_context
+from zenml.pipelines import get_pipeline_context
+from zenml.pipelines import pipeline
+from zenml.steps import step
+from zenml.steps.step_context import get_step_context
 from zenml.steps.utils import log_step_metadata
 from zenml.entrypoints import entrypoint
