+++ src/zenml/artifacts/artifact_config.py
@@ -21,7 +21,7 @@ from zenml.enums import ModelStages
 from zenml.exceptions import StepContextError
 from zenml.logger import get_logger
 from zenml.metadata.metadata_types import MetadataType
-from zenml.new.steps.step_context import get_step_context
+from zenml.steps.step_context import get_step_context
 
 if TYPE_CHECKING:
     from zenml.model.model import Model
