+++ src/zenml/artifacts/utils.py
@@ -50,8 +50,8 @@ from zenml.models import (
     StepRunResponse,
     StepRunUpdate,
 )
-from zenml.new.steps.step_context import get_step_context
 from zenml.stack import StackComponent
+from zenml.steps.step_context import get_step_context
 from zenml.utils import source_utils
 from zenml.utils.yaml_utils import read_yaml, write_yaml
