+++ src/zenml/hooks/hook_validators.py
@@ -43,9 +43,6 @@ def resolve_and_validate_hook(hook: "HookSpecification") -> Source:
     if not callable(func):
         raise ValueError(f"{func} is not a valid function.")
 
-    from zenml.new.steps.step_context import StepContext
-    from zenml.steps.base_parameters import BaseParameters
-
     sig = inspect.getfullargspec(inspect.unwrap(func))
     sig_annotations = sig.annotations
     if "return" in sig_annotations:
@@ -62,11 +59,7 @@ def resolve_and_validate_hook(hook: "HookSpecification") -> Source:
         seen_annotations = set()
         for annotation in annotations:
             if annotation:
-                if annotation not in (
-                    BaseException,
-                    BaseParameters,
-                    StepContext,
-                ):
+                if annotation not in (BaseException,):
                     raise ValueError(
                         "Hook arguments must be of type `BaseException`, not "
                         f"`{annotation}`."
