+++ src/zenml/zen_server/template_execution/utils.py
@@ -35,8 +35,8 @@ from zenml.models import (
     RunTemplateResponse,
     StackResponse,
 )
-from zenml.new.pipelines.build_utils import compute_stack_checksum
-from zenml.new.pipelines.run_utils import (
+from zenml.pipelines.build_utils import compute_stack_checksum
+from zenml.pipelines.run_utils import (
     create_placeholder_run,
     get_default_run_name,
     validate_run_config_is_runnable_from_server,
